+++ src/app/story-details/story-details.component.ts
@@ -1,19 +1,21 @@
 import { Component, OnInit } from '@angular/core';
 import { StoryService } from '../services/story.service';
 import { Story } from '../story/story';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Epic } from '../epic/epic';
 import { EpicService } from '../services/epic.service';
 import { Task } from '../task/task';
 import { TaskService } from '../services/task.service';
 import { MatTableDataSource } from '@angular/material';
-
+import { DeleteConfirmComponent } from '../delete-confirm/delete-confirm.component';
+import { MatDialog } from '@angular/material';
 @Component({
   selector: 'app-story-details',
   templateUrl: './story-details.component.html',
   styleUrls: ['./story-details.component.css']
 })
 export class StoryDetailsComponent implements OnInit {
+  result = false;
   id: number;
   private sub: any;
   story: Story;
@@ -24,7 +26,8 @@ export class StoryDetailsComponent implements OnInit {
   displayedColumns = ['Id', 'Name', 'EstTime', 'Add'];
   dataSource: any;
   constructor(private route: ActivatedRoute, private storyService: StoryService,
-              private epicService: EpicService, private taskService: TaskService) { }
+              private epicService: EpicService, private taskService: TaskService,
+              private dialog: MatDialog, public router: Router) { }
 
   ngOnInit() {
     this.sub = this.route.params.subscribe(params => {
@@ -41,12 +44,16 @@ export class StoryDetailsComponent implements OnInit {
       .subscribe( story => this.story = story,
         error => console.log('Error: ', error),
         () => this.taskService.getByStoryId(this.story.id)
-          .subscribe( tasks => this.tasks = tasks,
+          .subscribe( assignedTasks => this.assignedTasks = assignedTasks,
             error => console.log('Error: ', error),
-            () => this.epicService.getById(this.story.epicId)
-              .subscribe( epic => this.epic = epic,
+            () => this.taskService.getAll()
+              .subscribe(tasks => this.tasks = tasks,
                 error => console.log('Error: ', error),
-                () => this.dataSource = new MatTableDataSource<Task>(this.tasks)
+                () => this.epicService.getById(this.story.epicId)
+                  .subscribe( epic => this.epic = epic,
+                    error => console.log('Error: ', error),
+                    () => this.dataSource = new MatTableDataSource<Task>(this.tasks)
+                  )
               )
           )
       );
@@ -101,10 +108,32 @@ export class StoryDetailsComponent implements OnInit {
     console.log(this.checkedTasks);
     for (const task of this.checkedTasks) {
       task.storyId = this.story.id;
+      this.taskService.update(task)
+        .subscribe();
     }
   }
 
+  deleteDialog() {
+    const dialogRef = this.dialog.open(DeleteConfirmComponent, {
+      height: '150px',
+      width: '300px',
+      data: { type: 'Story', name: this.story.name}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result !== undefined) {
+        this.result = true;
+      }
+    }, error => console.log('Error: ', error),
+    () => this.delete());
+  }
+
   delete() {
-    this.storyService.delete(this.story.id);
+    if (this.result === true) {
+      this.storyService.delete(this.story.id)
+        .subscribe(empty => this.result = false,
+          error => console.log('Error: ', error),
+          () => this.router.navigate(['../../../backlog'])
+        );
+    }
   }
 }
